examples/basic-chat-agent/requirements.txt
#AgentArc OnChain Agent Dependencies

# Core dependencies
agentarc>=0.2.0
coinbase-agentkit>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web3 dependencies (included with agentarc)
web3>=6.0.0
eth-abi>=4.0.0

# FastAPI server with event streaming
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Optional: For development
pytest>=7.0.0
black>=23.0.0
